Fix Issue of Mobile Navigation
While testing the responsiveness, found one bug:

If the mobile nav is once opened and then closed, streching the browser window is not making any changes to navitems, due to which navbar was not visible.

Fixed this issue by adding additional classes and removing it using classList method in javascript.